android/core/statistics/src/main/java/ca/josephroque/bowlingcompanion/core/statistics/trackable/firstroll/HeadPinsSparedStatistic.kt
package ca.josephroque.bowlingcompanion.core.statistics.trackable.firstroll

import ca.josephroque.bowlingcompanion.core.model.TrackableFilter
import ca.josephroque.bowlingcompanion.core.model.TrackableFrame
import ca.josephroque.bowlingcompanion.core.model.arePinsCleared
import ca.josephroque.bowlingcompanion.core.model.isHeadPin
import ca.josephroque.bowlingcompanion.core.model.isHeadPin2
import ca.josephroque.bowlingcompanion.core.statistics.PreferredTrendDirection
import ca.josephroque.bowlingcompanion.core.statistics.R
import ca.josephroque.bowlingcompanion.core.statistics.StatisticCategory
import ca.josephroque.bowlingcompanion.core.statistics.StatisticID
import ca.josephroque.bowlingcompanion.core.statistics.TrackablePerFrameConfiguration
import ca.josephroque.bowlingcompanion.core.statistics.TrackablePerSecondRoll
import ca.josephroque.bowlingcompanion.core.statistics.interfaces.SecondRollStatistic

data class HeadPinsSparedStatistic(
	var headPins: Int = 0,
	var headPinsSpared: Int = 0,
) : TrackablePerSecondRoll, SecondRollStatistic {
	override val id = StatisticID.HEAD_PINS_SPARED
	override val denominatorTitleResourceId: Int = R.string.statistic_title_head_pins
	override val category = StatisticCategory.HEAD_PINS
	override val isEligibleForNewLabel = false
	override val preferredTrendDirection = PreferredTrendDirection.UPWARDS
	override fun emptyClone() = HeadPinsSparedStatistic()

	override var denominator: Int
		get() = headPins
		set(value) {
			headPins = value
		}

	override var numerator: Int
		get() = headPinsSpared
		set(value) {
			headPinsSpared = value
		}

	override fun adjustByFirstRollFollowedBySecondRoll(
		firstRoll: TrackableFrame.Roll,
		secondRoll: TrackableFrame.Roll,
		configuration: TrackablePerFrameConfiguration,
	) {
		if (firstRoll.pinsDowned.isHeadPin() || (
				configuration.countHeadPin2AsHeadPin &&
					firstRoll.pinsDowned.isHeadPin2()
				)
		) {
			headPins++

			if (secondRoll.pinsDowned.plus(firstRoll.pinsDowned).arePinsCleared()) {
				headPinsSpared++
			}
		}
	}

	override fun supportsSource(source: TrackableFilter.Source): Boolean = when (source) {
		is TrackableFilter.Source.Bowler -> true
		is TrackableFilter.Source.League -> true
		is TrackableFilter.Source.Series -> true
		is TrackableFilter.Source.Game -> true
	}
}
